// client/src/redux/action-types/auth.types.ts
export enum ActionType {
  DEPOSIT = "deposit",
  WITHDRAW = "withdraw",
  BANKRUPT = "bankrupt",
  REGISTER_USER = "REGISTER_USER",
  REGISTER_USER_FAIL = "REGISTER_USER_FAIL",
  LOGIN_USER = "LOGIN_USER",
  LOGIN_USER_FAIL = "LOGIN_USER_FAIL",
  USER_LOADED = "USER_LOADED",
  USER_LOADED_FAIL = "USER_LOADED_FAIL",
  CLEAR_ERROR = "CLEAR_ERROR",
  UPDATE_USER = "UPDATE_USER",
  DELETE_USER = "DELETE_USER",
  GET_SPECIFIC_USER = "GET_SPECIFIC_USER",
  USER_LOADING = "USER_LOADING",
}